requirements.txt
alpaca-py==0.38.0
requests==2.32.3
pandas==2.2.3
numpy==2.2.4
python-dotenv==1.1.0
schedule==1.2.2
rich==14.0.0
websocket-client==1.8.0
